organize folder structure: spec/ and ir/
