+++ src/ebpf_verifier.hpp
@@ -5,6 +5,6 @@
 #include "config.hpp"
 #include "crab_verifier.hpp"
 #include "elf_loader.hpp"
+#include "ir/program.hpp"
 #include "ir/unmarshal.hpp"
 #include "platform.hpp"
-#include "program.hpp"
