+++ src/linux/gpl/spec_type_descriptors.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "ebpf_base.h"
+#include "spec/ebpf_base.h"
 
 namespace prevail {
 constexpr int NMAPS = 64;
