+++ src/linux/linux_platform.hpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 #pragma once
 
-#include "helpers.hpp"
+#include "spec/function_prototypes.hpp"
 
 namespace prevail {
 EbpfHelperPrototype get_helper_prototype_linux(int32_t n);
