+++ src/spec/vm_isa.hpp
@@ -1,6 +1,7 @@
 // Copyright (c) Prevail Verifier contributors.
 // SPDX-License-Identifier: MIT
 #pragma once
+
 #include <cinttypes>
 #include <tuple>
