+++ src/string_constraints.hpp
@@ -6,7 +6,6 @@
 #include <optional>
 #include <set>
 #include <string>
-#include <tuple>
 #include <vector>
 
 #include "arith/linear_constraint.hpp"
